Erste kontrollierte Bewegung!

+++ trunk/CM-Bot/include/dynamixel.h
@@ -13,7 +13,7 @@
 #define DNX_BRDCAST_ID 0xFE
 
 byte DNX_getChecksum(byte*, byte);
-void DNX_setAngle(byte, uint16_t);
+void DNX_setAngle(byte, double);
 void DNX_setId(byte, byte);
 void DNX_setSpeed(byte, byte);
 void DNX_setLed(byte, byte);
